backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
sqlalchemy==2.0.36
pydantic==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pytest==8.3.0
httpx==0.27.0
python-dotenv==1.0.1
mcp>=1.0.0
